Add folders
Adding folders for more organization.

+++ category.php
@@ -1,6 +1,6 @@
 <?php
-require_once('classUser.php');
-require_once('functionsCategory.php');
+require_once('class/classUser.php');
+require_once('functions/functionsCategory.php');
 session_start();
 $user = $_SESSION['user'];
 $message = "";
@@ -15,53 +15,53 @@ if (empty($_SESSION) || $user->get_role() != 'Administrador') {
 <!DOCTYPE html>
 <html lang="en">
 
-    <head>
-        <link rel="shortcut icon" href="https://www.utn.ac.cr/misc/favicon.ico" type="image/vnd.microsoft.icon" />
-    </head>
+<head>
+    <link rel="shortcut icon" href="https://www.utn.ac.cr/misc/favicon.ico" type="image/vnd.microsoft.icon" />
+</head>
 
-    <body>
+<body>
 
-        <div class="container">
-            <?php require_once('header.php') ?>
+    <div class="container">
+        <?php require_once('header.php') ?>
 
-            <section>
-                <h3 style="text-align: center;">Category</h3>
-                <div class="container">
-                    <div class="row">
-                        <div class="four columns">
-                            <a href="/createCategory.php" class="button button-primary">Create</a>
-                        </div>
-                        <div class="three columns">
-                            <p style="color: red;"><?php echo $message; ?></p>
-                        </div>
-                        <div class="five columns">
-                            <div class="u-pull-right">
-                                <form action="category.php" method="GET">
-                                    <input type="text" placeholder="Search" name="search" title="Search for name">
-                                    <button type="submit"><i class="fas fa-search" style="color:grey"></i></button>
-                                </form>
-                            </div>
+        <section>
+            <h3 style="text-align: center;">Category</h3>
+            <div class="container">
+                <div class="row">
+                    <div class="four columns">
+                        <a href="/createCategory.php" class="button button-primary">Create</a>
+                    </div>
+                    <div class="three columns">
+                        <p style="color: red;"><?php echo $message; ?></p>
+                    </div>
+                    <div class="five columns">
+                        <div class="u-pull-right">
+                            <form action="category.php" method="GET">
+                                <input type="text" placeholder="Search" name="search" title="Search for name">
+                                <button type="submit"><i class="fas fa-search" style="color:grey"></i></button>
+                            </form>
                         </div>
                     </div>
-                    <table class="u-full-width">
-                        <thead>
-                            <tr>
-                                <th>Id</th>
-                                <th>Category</th>
-                                <th>Parent</th>
-                                <th>Utility</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <?php
-                        loadTable(); ?>
-                        </tbody>
-                    </table>
                 </div>
-            </section>
-        </div>
+                <table class="u-full-width">
+                    <thead>
+                        <tr>
+                            <th>Id</th>
+                            <th>Category</th>
+                            <th>Parent</th>
+                            <th>Utility</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <?php
+                        loadTable(); ?>
+                    </tbody>
+                </table>
+            </div>
+        </section>
+    </div>
 
-    </body>
+</body>
 
 </html>
 
@@ -71,17 +71,16 @@ function loadTable()
 {
     $categories = searchCategories($_GET && !empty($_REQUEST['search']) ? $_REQUEST['search'] : "");
     foreach ($categories as $category) { ?>
-<tr>
-    <td> <?php echo $category->get_id(); ?> </td>
-    <td> <?php echo $category->get_name(); ?> </td>
-    <td> <?php echo $category->get_parent(); ?> </td>
-    <td>
-        <a href="editCategory.php?id=<?php echo $category->get_id(); ?>">Edit</a> |
-        <a href="/CRUDCategories.php?action=delete&id=<?php echo $category->get_id(); ?>"
-            onclick="return confirm('Are you sure?');">Delete</a>
-    </td>
+        <tr>
+            <td> <?php echo $category->get_id(); ?> </td>
+            <td> <?php echo $category->get_name(); ?> </td>
+            <td> <?php echo $category->get_parent(); ?> </td>
+            <td>
+                <a href="editCategory.php?id=<?php echo $category->get_id(); ?>">Edit</a> |
+                <a href="/CRUD/CRUDCategories.php?action=delete&id=<?php echo $category->get_id(); ?>" onclick="return confirm('Are you sure?');">Delete</a>
+            </td>
 
-</tr>
+        </tr>
 
 <?php }
 }
