+++ product.php
@@ -1,6 +1,6 @@
 <?php
-require_once('classUser.php');
-require_once('functionsproduct.php');
+require_once('class/classUser.php');
+require_once('functions/functionsProduct.php');
 session_start();
 $user = $_SESSION['user'];
 if (empty($_SESSION) || $user->get_role() != 'Administrador') {
@@ -11,49 +11,49 @@ if (empty($_SESSION) || $user->get_role() != 'Administrador') {
 <!DOCTYPE html>
 <html lang="en">
 
-    <head>
-        <link rel="shortcut icon" href="https://www.utn.ac.cr/misc/favicon.ico" type="image/vnd.microsoft.icon" />
-    </head>
+<head>
+    <link rel="shortcut icon" href="https://www.utn.ac.cr/misc/favicon.ico" type="image/vnd.microsoft.icon" />
+</head>
 
-    <body>
+<body>
 
-        <div class="container">
-            <?php require_once('header.php') ?>
+    <div class="container">
+        <?php require_once('header.php') ?>
 
-            <section>
-                <h3 style="text-align: center;">Product</h3>
-                <div class="container">
-                    <div class="row">
-                        <div class="six columns">
-                            <a href="createProduct.php" class="button button-primary">Create</a>
-                        </div>
-                        <div class="six columns">
-                            <div class="u-pull-right">
-                                <form action="/product.php" method="GET">
-                                    <input type="text" placeholder="Search" name="search" title="Search for name">
-                                    <button type="submit"><i class="fas fa-search" style="color:grey"></i></button>
-                                </form>
-                            </div>
+        <section>
+            <h3 style="text-align: center;">Product</h3>
+            <div class="container">
+                <div class="row">
+                    <div class="six columns">
+                        <a href="createProduct.php" class="button button-primary">Create</a>
+                    </div>
+                    <div class="six columns">
+                        <div class="u-pull-right">
+                            <form action="/product.php" method="GET">
+                                <input type="text" placeholder="Search" name="search" title="Search for name">
+                                <button type="submit"><i class="fas fa-search" style="color:grey"></i></button>
+                            </form>
                         </div>
                     </div>
-                    <table class="u-full-width">
-                        <thead>
-                            <tr>
-                                <th>SKU</th>
-                                <th>Name</th>
-                                <th>Category</th>
-                                <th>Utility</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            <?php loadTable(); ?>
-                        </tbody>
-                    </table>
                 </div>
-            </section>
-        </div>
+                <table class="u-full-width">
+                    <thead>
+                        <tr>
+                            <th>SKU</th>
+                            <th>Name</th>
+                            <th>Category</th>
+                            <th>Utility</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <?php loadTable(); ?>
+                    </tbody>
+                </table>
+            </div>
+        </section>
+    </div>
 
-    </body>
+</body>
 
 </html>
 
@@ -62,23 +62,22 @@ if (empty($_SESSION) || $user->get_role() != 'Administrador') {
 
 function loadTable()
 {
-    include_once('functionsCategory.php');
+    include_once('functions/functionsCategory.php');
     $products = searchProduct($_GET ? $_REQUEST['search'] : "");
 
     foreach ($products as $product) {
         $category = categoryById($product->getIdCategory());
 ?>
-<tr>
-    <td> <?php echo $product->getSKU(); ?> </td>
-    <td> <?php echo $product->getName(); ?> </td>
-    <td> <?php echo $category->get_name(); ?> </td>
-    <td>
-        <a href="editProduct.php?id=<?php echo $product->getId(); ?>">Edit</a> |
-        <a href="/CRUDProducts.php?action=delete&id=<?php echo $product->getId(); ?>"
-            onclick="return confirm('Are you sure?')">Delete</a>
-    </td>
+        <tr>
+            <td> <?php echo $product->getSKU(); ?> </td>
+            <td> <?php echo $product->getName(); ?> </td>
+            <td> <?php echo $category->get_name(); ?> </td>
+            <td>
+                <a href="editProduct.php?id=<?php echo $product->getId(); ?>">Edit</a> |
+                <a href="/CRUD/CRUDProducts.php?action=delete&id=<?php echo $product->getId(); ?>" onclick="return confirm('Are you sure?')">Delete</a>
+            </td>
 
-</tr>
+        </tr>
 
 <?php }
 } ?>
